requirements.txt
asgiref==3.8.1
Django
django-rest-framework==0.1.0
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.12.2
